apiserver-watcher: use a lockfile.
We eventually want to rename this pod (since it's in the wrong
namespace), but before we can do that, we have to handle multiple
instances of the process. This is because the kubelet often handles
static pod updates by creating the new one before deleting the old one.

So, write a lockfile. Fortunately, we're hostPID, so we can actually
detect stale lockfiles. Whew.